requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==5.2.0
mypy-extensions==0.4.3
nltk==3.7
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyee==8.2.2
pyflakes==2.4.0
pyparsing==3.0.7
python-dotenv==0.19.1
regex==2021.10.23
requests==2.26.0
slack-sdk==3.14.1
slackeventsapi==3.0.0
tomli==1.2.2
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.0
